# swagger.json
{
  "swagger": "2.0",
  "info": {
    "title": "Typoteka API",
    "description": "Api docs for typoteka personal blog",
    "version": "1.0.0"
  },
  "host": "localhost:3000",
  "basePath": "/api",
  "tags": [
    {
      "name": "Articles",
      "description": "Articles endpoints"
    },
    {
      "name": "Categories",
      "description": "Categories endpoints"
    },
    {
      "name": "Comments",
      "description": "Comments endpoints"
    },
    {
      "name": "My",
      "description": "Endpoints for personal entities"
    },
    {
      "name": "Search",
      "description": "Search endpoints"
    },
    {
      "name": "Users",
      "description": "Users endpoints"
    }
  ],
  "schemes": [
    "http"
  ],
  "paths": {
    "/articles": {
      "get": {
        "tags": [
          "Articles"
        ],
        "summary": "Get all articles",
        "description": "",
        "parameters": [
          {
            "name": "offset",
            "in": "query",
            "type": "string"
          },
          {
            "name": "limit",
            "in": "query",
            "type": "string"
          },
          {
            "name": "fromCategoryId",
            "in": "query",
            "type": "string"
          }
        ],
        "responses": {}
      },
      "post": {
        "tags": [
          "Articles"
        ],
        "summary": "Create new article",
        "description": "",
        "parameters": [],
        "responses": {}
      }
    },
    "/articles/commented": {
      "get": {
        "tags": [
          "Articles"
        ],
        "summary": "Get most commented articles",
        "description": "",
        "parameters": [
          {
            "name": "limit",
            "in": "query",
            "type": "string"
          }
        ],
        "responses": {}
      }
    },
    "/articles/{:id}": {
      "get": {
        "tags": [
          "Articles"
        ],
        "summary": "Get article by id",
        "description": "",
        "parameters": [
          {
            "name": ":id",
            "in": "path",
            "required": true,
            "type": "string"
          }
        ],
        "responses": {}
      },
      "put": {
        "tags": [
          "Articles"
        ],
        "summary": "Edit article by id",
        "description": "",
        "parameters": [
          {
            "name": ":id",
            "in": "path",
            "required": true,
            "type": "string"
          }
        ],
        "responses": {}
      },
      "delete": {
        "tags": [
          "Articles"
        ],
        "summary": "Delete article by id",
        "description": "",
        "parameters": [
          {
            "name": ":id",
            "in": "path",
            "required": true,
            "type": "string"
          }
        ],
        "responses": {}
      }
    },
    "/articles/{:id}/comments": {
      "post": {
        "tags": [
          "Articles"
        ],
        "summary": "Create comment for article",
        "description": "",
        "parameters": [
          {
            "name": ":id",
            "in": "path",
            "required": true,
            "type": "string"
          }
        ],
        "responses": {}
      }
    },
    "/articles/{:articleId}/comments/{:commentId}": {
      "delete": {
        "tags": [
          "Articles"
        ],
        "summary": "Delete comment for article",
        "description": "",
        "parameters": [
          {
            "name": ":articleId",
            "in": "path",
            "required": true,
            "type": "string"
          },
          {
            "name": ":commentId",
            "in": "path",
            "required": true,
            "type": "string"
          }
        ],
        "responses": {}
      }
    },
    "/categories": {
      "get": {
        "tags": [
          "Categories"
        ],
        "summary": "Get all categories",
        "description": "",
        "parameters": [],
        "responses": {}
      },
      "post": {
        "tags": [
          "Categories"
        ],
        "summary": "Create new category",
        "description": "",
        "parameters": [],
        "responses": {}
      }
    },
    "/categories/{:id}": {
      "put": {
        "tags": [
          "Categories"
        ],
        "summary": "Edit category by id",
        "description": "",
        "parameters": [
          {
            "name": ":id",
            "in": "path",
            "required": true,
            "type": "string"
          }
        ],
        "responses": {}
      },
      "delete": {
        "tags": [
          "Categories"
        ],
        "summary": "Delete category by id",
        "description": "",
        "parameters": [
          {
            "name": ":id",
            "in": "path",
            "required": true,
            "type": "string"
          }
        ],
        "responses": {}
      }
    },
    "/comments/latest": {
      "get": {
        "tags": [
          "Comments"
        ],
        "summary": "Get latest comments",
        "description": "",
        "parameters": [
          {
            "name": "limit",
            "in": "query",
            "type": "string"
          }
        ],
        "responses": {}
      }
    },
    "/my": {
      "get": {
        "tags": [
          "My"
        ],
        "summary": "Get user",
        "description": "",
        "parameters": [],
        "responses": {}
      }
    },
    "/my/comments": {
      "get": {
        "tags": [
          "My"
        ],
        "summary": "Get user",
        "description": "",
        "parameters": [],
        "responses": {}
      }
    },
    "/search": {
      "get": {
        "tags": [
          "Search"
        ],
        "summary": "Find article",
        "description": "",
        "parameters": [
          {
            "name": "query",
            "in": "query",
            "type": "string"
          }
        ],
        "responses": {}
      }
    },
    "/users": {
      "post": {
        "tags": [
          "Users"
        ],
        "summary": "Create user",
        "description": "",
        "parameters": [],
        "responses": {}
      }
    },
    "/users/auth": {
      "post": {
        "tags": [
          "Users"
        ],
        "summary": "Auth with JWT token",
        "description": "",
        "parameters": [
          {
            "name": "body",
            "in": "body",
            "schema": {
              "type": "object",
              "properties": {
                "email": {
                  "example": "any"
                },
                "password": {
                  "example": "any"
                }
              }
            }
          }
        ],
        "responses": {}
      }
    },
    "/users/refresh": {
      "post": {
        "tags": [
          "Users"
        ],
        "summary": "Refresh JWT tokens",
        "description": "",
        "parameters": [
          {
            "name": "body",
            "in": "body",
            "schema": {
              "type": "object",
              "properties": {
                "token": {
                  "example": "any"
                }
              }
            }
          }
        ],
        "responses": {}
      }
    }
  }
}